feat: Products CRUD e correções
- CRUD de produtos
- ProductFormModal com validação Zod e máscaras
- Corrige imports TypeScript para verbatimModuleSyntax
- Simplifica schemas Zod (phone, segment_id, status)
- Remove interceptor loop em api.ts com verificação /auth/
- Atualiza docker-compose.yml para Docker Compose v2

+++ backend/app/Http/Controllers/API/Dashboard/DashboardController.php
@@ -193,7 +193,7 @@ class DashboardController extends Controller
             )
             ->whereNotNull('opportunities.assigned_to')
             ->orderBy('opportunities.created_at', 'desc')
-            ->limit(ceil($limit / 2))
+            ->limit((int) ceil($limit / 2))
             ->get();
 
         // Buscar últimos clientes cadastrados
@@ -206,7 +206,7 @@ class DashboardController extends Controller
                 'customers.created_at'
             )
             ->orderBy('customers.created_at', 'desc')
-            ->limit(ceil($limit / 2))
+            ->limit((int) ceil($limit / 2))
             ->get();
 
         // Combinar e ordenar por data
